refactor(subagents): extract launcher/env into subagent-launcher.ts
Move subagent process-launch resolution out of manager.ts into a new
src/agent/subagents/subagent-launcher.ts: resolveSubagentLauncher,
resolveSubagentWorkingDirectory, composeSubagentChildEnv,
resolveSubagentInheritedPrimaryRoot, buildInheritedChannelContextPayload.

Pure move, no behavior change. SubagentMemoryScope (used by both the leaf
and manager) moves to the subagents index so the new module and manager can
share it without a cycle. Importers (manager + three tests) import from the
new module / index directly. manager.ts baseline ratcheted 1357 -> 1117.
